% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblogistic.R
\name{find_tuning_constantMqle}
\alias{find_tuning_constantMqle}
\title{Find tuning constant for robust logistic (Mqle) with Tukey's weight}
\usage{
find_tuning_constantMqle(
  x,
  beta,
  c_min,
  c_max,
  eff = 0.95,
  inconsistent = FALSE
)
}
\arguments{
\item{x}{design matrix}

\item{beta}{parameter}

\item{c_min}{minimum tuning parameter value}

\item{c_max}{maximum tuning parameter value}

\item{eff}{level of efficiency (95\% by default)}

\item{inconsistent}{efficiency at the inconsistent estimator?}
}
\description{
Find tuning constant in the robust logistic with Tukey's weight
for a given desired level of efficiency depending on the design and
the true parameter beta.
}
