% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JINIpaper-package.R
\docType{package}
\name{JINIpaper}
\alias{JINIpaper-package}
\alias{JINIpaper}
\title{JINIpaper: Code for the paper "A Flexible Bias Correction Method based on Inconsistent Estimators"}
\description{
This package contains the code to replicate the numerical studies presented in the paper "A Flexible Bias Correction Method based on Inconsistent Estimators".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SMAC-Group/JINIpaper}
  \item Report bugs at \url{https://github.com/SMAC-Group/JINIpaper/issues/}
}

}
\author{
\strong{Maintainer}: Samuel Orso \email{Samuel.Orso@unige.ch} [copyright holder]

Authors:
\itemize{
  \item Stéphane Guerrier [contributor]
  \item Yuming Zhang [contributor]
  \item Yanyuan Ma
  \item Mucyo Karemera
  \item Maria-Pia Victoria-Feser
}

Other contributors:
\itemize{
  \item R Core Team [copyright holder]
  \item `MASS` Package Team [copyright holder]
  \item `betareg` Package Team [copyright holder]
}

}
\keyword{internal}
