% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_mqle}
\alias{logistic_mqle}
\title{MQLE for logistic regression}
\usage{
logistic_mqle(y, x, c)
}
\arguments{
\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}
}
\description{
MQLE for logistic regression
}
