% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_wmle}
\alias{logistic_wmle}
\title{MLE for logistic regression with misclassified responses}
\usage{
logistic_wmle(y, x, c)
}
\arguments{
\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}
}
\description{
MLE for logistic regression with misclassified responses
}
