% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_wmle_stocapp}
\alias{logistic_wmle_stocapp}
\title{Stochastic approximation for robust logistic regression with inconsistent initial estimator with Tukey's weights}
\usage{
logistic_wmle_stocapp(
  x,
  start,
  c = 4.685061,
  maxit = 10000L,
  tol = 1e-05,
  verbose = FALSE,
  seed = 321L
)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{start}{an inconsistent estimator (also used as starting values)}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}

\item{maxit}{max number of iteration for IRWLS}

\item{tol}{tolerance for stopping criterion}

\item{verbose}{print info}

\item{seed}{for random number generator}

\item{H}{number of estimators for Monte Carlo approximation}
}
\description{
Stochastic approximation for robust logistic regression with inconsistent initial estimator with Tukey's weights
}
