% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto_vglm_ib}
\alias{pareto_vglm_ib}
\title{Iterative bootstrap for robust Pareto regression}
\usage{
pareto_vglm_ib(
  x,
  thetastart,
  H = 200,
  maxit = 200,
  tol = 1e-07,
  verbose = FALSE,
  seed = 321
)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{thetastart}{an inconsistent estimator (also used as starting values)}

\item{H}{number of estimators for Monte Carlo approximation}

\item{maxit}{max number of iteration for IRWLS}

\item{tol}{tolerance for stopping criterion}

\item{verbose}{print info}

\item{seed}{for random number generator}

\item{c}{tuning parameter for Huber's weight (default value is 1.345)}
}
\description{
Iterative bootstrap for robust Pareto regression
}
