% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paretoMle}
\alias{paretoMle}
\title{Pareto regression (MLE)}
\usage{
paretoMle(y, x, start, maxit = 200L, tol = 1e-07, verbose = FALSE)
}
\arguments{
\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{start}{is a p-vector of parameters (starting values)}

\item{maxit}{max number of iteration for IRWLS}

\item{tol}{tolerance for stopping criterion}

\item{verbose}{print info}
}
\description{
Pareto regression (MLE)
}
