% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{paretoWmle1_ib}
\alias{paretoWmle1_ib}
\title{Iterative bootstrap for robust Pareto regression with inconsistent initial estimator with Tukey's weights}
\usage{
paretoWmle1_ib(
  x,
  thetastart,
  c = 4.685061,
  H = 200,
  maxit = 200,
  tol = 1e-07,
  verbose = FALSE,
  seed = 321
)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{thetastart}{an inconsistent estimator (also used as starting values)}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}

\item{H}{number of estimators for Monte Carlo approximation}

\item{maxit}{max number of iteration for IRWLS}

\item{tol}{tolerance for stopping criterion}

\item{verbose}{print info}

\item{seed}{for random number generator}
}
\description{
Iterative bootstrap for robust Pareto regression with inconsistent initial estimator with Tukey's weights
}
