% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paretoWmle1H}
\alias{paretoWmle1H}
\title{Robust Pareto regression initial estimator (inconsistent) with Huber's weights}
\usage{
paretoWmle1H(
  y,
  x,
  start,
  c = 1.345,
  maxit = 200L,
  tol = 1e-07,
  verbose = FALSE
)
}
\arguments{
\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Huber's weight (default value is 1.345)}

\item{maxit}{max number of iteration for IRWLS}

\item{tol}{tolerance for stopping criterion}

\item{verbose}{print info}

\item{beta}{a p-vector of parameter (starting values)}
}
\description{
Robust Pareto regression initial estimator (inconsistent) with Huber's weights
}
