% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paretoWmle_of}
\alias{paretoWmle_of}
\title{Robust Pareto regression initial estimator (inconsistent) with Tukey's weights}
\usage{
paretoWmle_of(start, y, x, c = 4.685061)
}
\arguments{
\item{start}{a p-vector of parameter (starting values)}

\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}
}
\description{
Robust Pareto regression initial estimator (inconsistent) with Tukey's weights
}
