% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_logistic}
\alias{r_logistic}
\title{Simulation of logistic regression}
\usage{
r_logistic(beta, x, seed)
}
\arguments{
\item{beta}{a p-vector of parameter (starting values)}

\item{x}{a n x p matrix of design}

\item{seed}{for random number generator}
}
\description{
Simulation of logistic regression
}
