% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_pareto}
\alias{r_pareto}
\title{Simulation of Pareto regression}
\usage{
r_pareto(beta, k, x, seed)
}
\arguments{
\item{beta}{a p-vector of parameter (starting values)}

\item{k}{minimum}

\item{x}{a n x p matrix of design}

\item{seed}{for random number generator}
}
\description{
Simulation of Pareto regression
}
