% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roblogisticMqleVar}
\alias{roblogisticMqleVar}
\title{Asymptotic variance of robust logistic regression estimator with Tukey's weights}
\usage{
roblogisticMqleVar(x, start, c = 4.685061)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Tukey's weights (default value is 4.685061)}

\item{beta}{a p-vector of parameter (starting values)}
}
\description{
Asymptotic variance of robust logistic regression estimator with Tukey's weights
}
