% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_wmle}
\alias{logistic_wmle}
\title{WMLE for logistic regression}
\usage{
logistic_wmle(y, x, c = 4.685061)
}
\arguments{
\item{y}{a vector of responses}

\item{x}{a n x p matrix of design}

\item{c}{tuning parameter for Tukey's weight (default value is 4.685061)}
}
\description{
WMLE for logistic regression
}
